requirements.txt
Django == 2.1.7
django_froala_editor == 2.9.3
django_richtextfield == 1.4.0
django_tinymce == 2.8.0
